requirements.txt
requests>=2.31.0
pandas>=2.2.0
duckdb>=0.10.0
streamlit>=1.32.0
plotly>=5.19.0
python-dotenv>=1.0.0